v13.5 fix

+++ WpfAppForModbus/Handlers/SensorHandler.cs
@@ -18,7 +18,7 @@ namespace WpfAppForModbus.Handlers
 
             int DecimalNumber = Convert.ToInt32(Registers[4], 16);
 
-            return DecimalNumber > 0 ? 4095 / DecimalNumber * 5 : 0;
+            return DecimalNumber > 0 ? 4095.0 / (double)DecimalNumber * 5.0 : 0;
         }
 
         public static double CountBar(string Result) {
@@ -26,7 +26,7 @@ namespace WpfAppForModbus.Handlers
 
             int DecimalNumber = Convert.ToInt32(Registers[4], 16);
 
-            return DecimalNumber > 0 ? 4095 / DecimalNumber * 5 : 0;
+            return DecimalNumber > 0 ? 4095.0 / (double)DecimalNumber * 5.0 : 0;
         }
 
         public static double CountWater(string Result) {
@@ -34,7 +34,7 @@ namespace WpfAppForModbus.Handlers
 
             int DecimalNumber = Convert.ToInt32(Registers[4], 16);
 
-            return DecimalNumber > 0 ? 4095 / DecimalNumber * 5 : 0;
+            return DecimalNumber > 0 ? 4095.0 / (double)DecimalNumber * 5.0 : 0;
         }
 
         public static double CountVoltage(string Result) {
@@ -42,7 +42,7 @@ namespace WpfAppForModbus.Handlers
 
             int DecimalNumber = Convert.ToInt32(Registers[4], 16);
 
-            return DecimalNumber > 0 ? 4095 / DecimalNumber * 5 : 0;
+            return DecimalNumber > 0 ? 4095.0 / (double)DecimalNumber * 5.0 : 0;
         }
     }
 }
